Rename _core.py -> _aliases.py
