+++ pyvista/core/_typing/__init__.py
@@ -1,4 +1,4 @@
 """Type aliases for type hints."""
 # flake8: noqa: F401
+from ._aliases import Array, BoundsLike, Matrix, Number, TransformLike, Vector
 from ._array_like import NumpyArray
-from ._core import Array, BoundsLike, Matrix, Number, TransformLike, Vector
